requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
geopy==2.4.1
pyswisseph==2.10.3.2
anthropic==0.26.0
stripe==9.12.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
weasyprint==62.3
